Extracted the CalculatorTester class
